requirements.txt
asgiref==3.5.2
automium==0.2.6
automium-web==0.1.1
backports.zoneinfo==0.2.1
captcha==0.4
certifi==2022.9.14
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
Django==4.1
django-debug-toolbar==3.6.0
django-ipware==4.0.2
django-ranged-response==0.2.0
django-recaptcha==3.0.0
django-simple-captcha==0.5.17
djangorestframework==3.14.0
fields==5.0.0
Flask==2.2.2
idna==3.4
importlib-metadata==4.12.0
itsdangerous==2.1.2
sqlparse==0.4.2
tzdata==2022.2
urllib3==1.26.12
views==0.3
Werkzeug==2.2.2
zipp==3.8.1
